client-hints/service-workers/intercept-request.https.html

<!DOCTYPE html>
<title>Client Hint/Service Worker interaction</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>
var SCRIPT='resources/intercept-request.js';
promise_test((t) =>
  service_worker_unregister_and_register(t, SCRIPT, SCRIPT)
  .then((r) => wait_for_state(t, r.installing, 'activated'))
  .then(() => {
      var popup_window = window.open("resources/foo.html");
      assert_not_equals(popup_window, null, "Popup windows not allowed?");
      popup_window.addEventListener('load', (e) => {
        t.step(()=>{assert_equals(popup_window.document.body.textContent, "PASS")});
        t.done();
      });
  })
, "Client Hint/Service Worker interaction: service workers succsefully receives hints from request");
</script>
