third_party/blink/web_tests/virtual/legacy-client-hints-no-fp-delegation/external/wpt/client-hints/service-workers/intercept-request.https-expected.txt

This is a testharness.js-based test.
FAIL Service workers succsefully receives hints from request assert_equals: expected "PASS" but got "FAIL"
Harness: the test ran to completion.
